+++ dbConnect/__init__.py
@@ -1,10 +1,12 @@
 """
 MySQL for Humans
 """
+from .dbConnect import DBConnect
+
 __description__ = 'MySQL for Humans'
 __author__ = "Emin Mastizada <emin@linux.com>"
-__version__ = '1.5.0'
+__version__ = '1.6.0'
 __license__ = "MPL 2.0"
 __help__ = """MySQL for Humans"""
 
-from .dbConnect import DBConnect
+__all__ = ['DBConnect', ]